correció 2

+++ divisio.py
@@ -1,5 +1,5 @@
 a=int(input("Escriu el teu dividend"))
-b=float(input("Esciu el teu divisor"))
+b=int(input("Esciu el teu divisor"))
 quocient=a//b
 residu=a%b
 print (f"Divisió: {a}/{b}")
